update private chat component, create chat component, and update file structure
