+++ frontend/src/__tests__/components/MessageForm.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { mount } from '@cypress/react';
-import MessageForm from '../../components/MessageForm';
+import MessageForm from '../../components/Chat/MessageForm';
 
 it('renders send button link', () => {
   mount(<MessageForm />);
