+++ frontend/src/components/Chat/MessageForm.tsx
@@ -2,8 +2,8 @@
  * Chat Message Text Input and Submit
  */
 import React, { useContext } from "react";
-import { SocketContext } from "../context/socket";
-import { useAppSelector } from '../app/hooks';
+import { SocketContext } from "../../context/socket";
+import { useAppSelector } from '../../app/hooks';
 
 
 const MessageForm: React.FC = () => {
