+++ frontend/src/components/Chat/MessagesDisplay/Message.tsx
@@ -6,7 +6,7 @@ import React, { memo } from "react";
 interface MessageProps {
   message: string;
   className: string;
-}
+};
 
 const Message: React.FC<MessageProps> = memo(({ message, className }) => {
   return (
