+++ frontend/src/components/Chat/MessagesDisplay/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './MessagesDisplay';
